Update twitter search to scope through #ubuntu hashtag
The #juju hashtag is a namespace which includes several topics, some of
which are not at all relevant to our namespace. For now, scope this
through the ubuntu hashtag as well, which will scope to mostly tweets
about juju relating to ubuntu. Sadly this will also negate any
OpenStack/Centos/Etc. based tweets about Juju that do not implicitly
include the ubuntu hashtag. Cest la vi